Add dotenv-based config
