+++ .env.example
@@ -9,3 +9,9 @@ COINGECKO_BASE_URL=
 LOG_LEVEL=INFO
 # File path for log output (default bot.log)
 LOG_FILE=bot.log
+# Default percent change that triggers an alert
+DEFAULT_THRESHOLD=0.1
+# Default subscription interval
+DEFAULT_INTERVAL=5m
+# How often prices are checked
+PRICE_CHECK_INTERVAL=30s
