+++ run.py
@@ -1,10 +1,8 @@
 #!/usr/bin/env python3
 import asyncio
-import json
 import logging
 import os
 import random
-import re
 import signal
 import time
 from collections import defaultdict, deque
@@ -20,7 +18,6 @@ import matplotlib
 import numpy as np
 from aiolimiter import AsyncLimiter
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
-from dotenv import load_dotenv
 from matplotlib import pyplot as plt
 from telegram import (
     Bot,
@@ -41,22 +38,20 @@ from telegram.ext import (
     filters,
 )
 
-matplotlib.use("Agg")
-
-load_dotenv()
-DB_FILE = os.getenv("DB_PATH", "subs.db")
-BOT_NAME = "PricePulseWatcherBot"
-DEFAULT_THRESHOLD = 0.1
-DEFAULT_INTERVAL = 300
-PRICE_CHECK_INTERVAL = 60
-
-# optional CoinGecko API key for higher rate limits
-COINGECKO_API_KEY = os.getenv("COINGECKO_API_KEY")
-# base API endpoint, override to use the pro API
-COINGECKO_BASE_URL = os.getenv("COINGECKO_BASE_URL", "https://api.coingecko.com/api/v3")
-COINGECKO_HEADERS = (
-    {"x-cg-pro-api-key": COINGECKO_API_KEY} if COINGECKO_API_KEY else None
+from pricepulsebot.config import (
+    BOT_NAME,
+    COINGECKO_BASE_URL,
+    COINGECKO_HEADERS,
+    DB_FILE,
+    DEFAULT_INTERVAL,
+    DEFAULT_THRESHOLD,
+    LOG_FILE,
+    LOG_LEVEL,
+    PRICE_CHECK_INTERVAL,
+    parse_duration,
 )
+
+matplotlib.use("Agg")
 
 # CoinGecko is the sole API provider
 
@@ -66,19 +61,6 @@ DOWN_ARROW = "\U0001f53b"  # down triangle
 ROCKET = "\U0001f680"  # rocket for big gains
 BOMB = "\U0001f4a3"  # bomb for big drops
 DEFAULT_ALERT_EMOJI = ROCKET
-
-
-def parse_duration(value: str) -> int:
-    """Return seconds for a duration string like '15m' or '1h'."""
-    # Accept a bare number of seconds or a value suffixed with d/h/m/s
-    if value.isdigit():
-        return int(value)
-    match = re.fullmatch(r"(\d+)([dhms])", value.lower())
-    if not match:
-        raise ValueError("invalid interval format")
-    num, unit = match.groups()
-    factor = {"d": 86400, "h": 3600, "m": 60, "s": 1}[unit]
-    return int(num) * factor
 
 
 def format_interval(seconds: int) -> str:
@@ -91,38 +73,6 @@ def format_interval(seconds: int) -> str:
     if seconds % 60 == 0:
         return f"{seconds // 60}m"
     return f"{seconds}s"
-
-
-def load_config(path: str = "config.json") -> None:
-    """Load defaults from a JSON config if present."""
-    # Values in the optional config file override hard coded defaults
-    if not os.path.isfile(path):
-        return
-    try:
-        with open(path) as fh:
-            data = json.load(fh)
-    except (OSError, json.JSONDecodeError) as exc:
-        logger.warning("failed to load config: %s", exc)
-        return
-
-    global DEFAULT_THRESHOLD, DEFAULT_INTERVAL, PRICE_CHECK_INTERVAL
-    if "default_threshold" in data:
-        try:
-            DEFAULT_THRESHOLD = float(data["default_threshold"])
-        except (TypeError, ValueError):
-            logger.warning("invalid default_threshold in config")
-    if "default_interval" in data:
-        try:
-            # parse_duration understands suffixes like h or m
-            DEFAULT_INTERVAL = parse_duration(str(data["default_interval"]))
-        except ValueError:
-            logger.warning("invalid default_interval in config")
-    if "price_check_interval" in data:
-        try:
-            # Frequency for periodic price checks
-            PRICE_CHECK_INTERVAL = parse_duration(str(data["price_check_interval"]))
-        except ValueError:
-            logger.warning("invalid price_check_interval in config")
 
 
 COINS = ["bitcoin", "ethereum", "litecoin", "dogecoin"]
@@ -286,13 +236,12 @@ async def fetch_top_coins() -> None:
         logger.error("error fetching top coins: %s", exc)
 
 
-LOG_FILE = os.getenv("LOG_FILE")
 _handlers = [logging.StreamHandler()]
 if LOG_FILE:
     _handlers.append(logging.FileHandler(LOG_FILE))
 
 logging.basicConfig(
-    level=os.getenv("LOG_LEVEL", "INFO").upper(),
+    level=LOG_LEVEL.upper(),
     handlers=_handlers,
     force=True,
 )
@@ -1408,8 +1357,7 @@ async def menu(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
 async def main() -> None:
     """Start the bot."""
     # Setup, register handlers and run until terminated
-    load_dotenv()
-    load_config()
+
     await init_db()
     await fetch_trending_coins()
     await fetch_top_coins()
